Fix incorrect format string.
  CC       config.lo
In file included from hercules.h:104,
                 from config.c:20:
config.c: In function 'configure_cpu':
config.c:1079:57: warning: format '%d' expects argument of type 'int', but argument 8 has type 'long int' [-Wformat=]
 1079 |             WRMSG( HHC00111, "I", _POSIX_THREAD_CPUTIME );
      |                                                         ^
msgenu.h:99:98: note: in definition of macro 'PFWRITEMSG'
   99 | #define PFWRITEMSG( pan, f, fmt, ... ) fwritemsg( __FILE__, __LINE__, __FUNCTION__, pan,   f,    fmt, ## __VA_ARGS__ )
      |                                                                                                  ^~~
msgenu.h:108:36: note: in expansion of macro 'PWRITEMSG'
  108 | #define  LOGMSG(      fmt,   ... ) PWRITEMSG(  WRMSG_NORMAL,       fmt,                 ## __VA_ARGS__ )
      |                                    ^~~~~~~~~
msgenu.h:110:36: note: in expansion of macro 'LOGMSG'
  110 | #define  WRMSG(     id, sev, ... ) LOGMSG(                      #id "%s " id "\n", sev, ## __VA_ARGS__ )
      |                                    ^~~~~~
config.c:1079:13: note: in expansion of macro 'WRMSG'
 1079 |             WRMSG( HHC00111, "I", _POSIX_THREAD_CPUTIME );
      |             ^~~~~

The `%d` comes from the definition of `HHC00111`:

`#define HHC00111 "Thread CPU Time IS available (_POSIX_THREAD_CPUTIME=%d)"`

and `_POSIX_THREAD_CPUTIME` comes from `<unistd.h>`:

`#define       _POSIX_THREAD_CPUTIME           200112L`

https://pubs.opengroup.org/onlinepubs/9699919799/ defines it as "shall either be -1 or 200809L"
So NetBSD is a bit behind with the value, but it is correct that it is a long int.

+++ config.c
@@ -1076,7 +1076,7 @@ int configure_cpu( int target_cpu )
         /* provided the _POSIX_THREAD_CPUTIME is supported.                       */
         pthread_getcpuclockid( sysblk.cputid[ target_cpu ], &sysblk.cpuclockid[ target_cpu ]);
         if (!sysblk.hhc_111_112)
-            WRMSG( HHC00111, "I", _POSIX_THREAD_CPUTIME );
+            WRMSG( HHC00111, "I", (long int)_POSIX_THREAD_CPUTIME );
 #else
         /* When not supported, we zero the cpuclockid, which will trigger a       */
         /* different approach to obtain the thread CPU time in clock.c            */
